v1.0.0

+++ bin/fx.zsh
@@ -1,9 +1,28 @@
 pglaunch() {
-  echo "test release"
+  echo
+
+  # check if docker is installed
+  which docker &>/dev/null || {
+    tput setaf 1
+    echo "Docker is not installed. Please install docker and try again."
+    tput sgr0
+    exit 1
+  }
+
+  # check if docker is running
+  [[ $(docker info &>/dev/null)? -ne 0 ]] && {
+    tput setaf 1
+    echo "Docker is not running. Please start docker and try again."
+    tput sgr0
+    exit 1
+  }
+
+  name=""
+  port=""
 
   while [[ "$#" -gt 0 ]]; do
     case "$1" in
-    # name for postgres container
+    # name for postgres container (default: current directory name)
     -n | --name)
       name="$2"
       if [[ -z "$name" ]]; then
@@ -13,7 +32,7 @@ pglaunch() {
       shift 2
       ;;
 
-    # port for postgres container
+    # port for postgres container (default: 5555)
     -p | --port)
       port="$2"
       if [[ -z "$port" ]]; then
@@ -30,48 +49,22 @@ pglaunch() {
     esac
   done
 
-  echo
+  # set defaults if not provided
+  [[ -z "$name" ]] && name="$(basename $(pwd))"
+  [[ -z "$port" ]] && port="5555"
 
-  # if name is not provided, assign current directory name
-  if [[ -z "$name" ]]; then
-    name="$(basename $(pwd))"
-  fi
-
-  # if port is not provided, prompt for it
-  if [[ -z "$port" ]]; then
-    tput setaf 3
-    echo -n "enter port for access (enter for default i.e. 5432): "
-    tput sgr0
-    read port
-    if [[ -z "$port" ]]; then
-      port="5432"
-    fi
-    echo
-  fi
-
-  # final variables with modifications if any
   name="$name-$(openssl rand -base64 3 | LC_CTYPE=C tr -dc 'a-zA-Z0-9' | head -c 4)"
   port="$port"
   pguname="postgres"
   pgupass="$(openssl rand -base64 12 | LC_CTYPE=C tr -dc 'a-zA-Z0-9' | head -c 12)"
   pgdname="postgres"
+
+  # check if postgres container is already running
   existing=$(docker ps -a | grep 5432 | awk '{print $1}' | head -n 1)
+  [[ -n "$existing" ]] && docker rm -f $existing &>/dev/null
 
-  if [[ -n "$existing" ]]; then
-    tput setaf 5
-    echo "> docker rm -f $existing"
-    tput sgr0
-    docker rm -f $existing
-    echo
-  fi
+  # run postgres container
+  docker run --name $name -e POSTGRES_USER=$pguname -e POSTGRES_PASSWORD=$pgupass -e POSTGRES_DB=$pgdname -p $port:5432 -d --rm postgres:alpine &>/dev/null
 
-  tput setaf 5
-  echo "> docker run --name $name -e POSTGRES_USER=$pguname -e POSTGRES_PASSWORD=$pgupass -e POSTGRES_DB=$pgdname -p $port:5432 -d --rm postgres:alpine"
-  tput sgr0
-  docker run --name $name -e POSTGRES_USER=$pguname -e POSTGRES_PASSWORD=$pgupass -e POSTGRES_DB=$pgdname -p $port:5432 -d --rm postgres:alpine
-  echo
-
-  tput setaf 14
-  echo "POSTGRES_URL=postgresql://$pguname:$pgupass@localhost:$port/$pgdname"
-  tput sgr0
+  echo "POSTGRES_URL=$(tput setaf 14)postgresql://$pguname:$pgupass@localhost:$port/$pgdname$(tput sgr0)"
 }
